move: フォルダ構成の整理 (Docker や Strapi と区別するために src ディレクトリに移行)
