+++ src/components/ui/toggle-group.tsx
@@ -6,7 +6,7 @@ import { VariantProps } from 'class-variance-authority';
 
 import { cn } from '@/lib/utils';
 import { toggleVariants } from '@/components/ui/toggle';
-import { CheckIcon } from '@/app/components/icons/material-symbols';
+import { CheckIcon } from '@/components/ui/icons';
 
 const ToggleGroupContext = React.createContext<
   VariantProps<typeof toggleVariants>
